feat(framework): Added runtime framework for RESTful API service
We use gin as the RESTful framework and use cobra as the command line framework to launch. During the life cycle, the process signal is listened to, but when the SIGHUP, SIGINT, SIGTERM, SIGQUIT, SIGUSR1, and SIGUSR2 signals are received, the RESTful service is stopped and the application framework is exited.