build(github-action): Split release action to develop merged publish and manual tag publish
Split release action for two way.
One way is action at develop merged. This way is just publish the develop image for test enviroment to test.
The another way is action at push 'v*' or 'V*' start tag. It is the manual publish tag. so it will make the release note in the github releases, and take the commit log to changelog.